requirements.txt
flask==2.2.2
supabase==0.5.8
google-generativeai==0.2.0
requests==2.28.1
